RSDK-9878: Move webService into web.go.
The webstream package is now always imported. Split up webstream package into C and not-C pieces to
accommodate this.
