+++ robot/web/web_notc.go
@@ -4,12 +4,10 @@ package web
 
 import (
 	"context"
-	"sync"
 
 	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/resource"
 	"go.viam.com/rdk/robot"
-	"go.viam.com/utils/rpc"
 )
 
 // New returns a new web service for the given robot.
@@ -27,25 +25,6 @@ func New(r robot.Robot, logger logging.Logger, opts ...Option) Service {
 		opts:      wOpts,
 	}
 	return webSvc
-}
-
-type webService struct {
-	resource.Named
-
-	mu         sync.Mutex
-	r          robot.Robot
-	rpcServer  rpc.Server
-	modServer  rpc.Server
-	services   map[resource.API]resource.APIResourceCollection[resource.Resource]
-	opts       options
-	addr       string
-	modAddr    string
-	logger     logging.Logger
-	cancelCtx  context.Context
-	cancelFunc func()
-	isRunning  bool
-	webWorkers sync.WaitGroup
-	modWorkers sync.WaitGroup
 }
 
 // Update updates the web service when the robot has changed.
